Made switching renderer 'more atomic'
When the Tcl setting 'renderer' is changed we need to destroy the old
renderer and create a new one. Though for technical reasons we don't do
that in the setting-changed Tcl-callback, instead we send a
SwitchRendererEvent and do the change when that events gets processed in
the main event loop handler, so when the Tcl callback is no longer
active.

Before this patch the above was implemented more or less like this:
    void Display::update(const Setting& setting) // triggers when 'renderer' setting gets changed
    {
        ...
        currentRenderer = renderSettings.getRenderer();
        getEventDistributor().distributeEvent(SwitchRendererEvent());
        ...
    }

Notice how the 'currentRenderer' variable is already updated before the
renderer has actually been changed.

There was some unfortunate interaction with the (old Tcl script based)
session management code:
* In `session_management.tcl` we have this code snippet:
    after realtime 0 { load_session "default_session" }
* Internally `load_session` loads (one or more) savestates.
* Loading a savestate creates a MSX machine which contains a VDP.
* Depending on the current renderer this creates a different `Rasterizer`.

But we now have the following combination:
* We have a pending `after realtime 0` event which will create a MSX
  machine.
* `currentRenderer` has already been changed, but the actual renderer
  switch is only scheduled yet (via `SwitchRendererEvent`).
* The `after realtime` event happens to get executed first. This already
  uses the updated value for `currentRenderer`, though at this point
  that's inconsistent with the actual renderer state.
* I've simplified and omitted a lot of details, but in the end this
  leads to a crash.

The fix is to rewrite the above code to:
    void Display::update(const Setting& setting) // triggers when 'renderer' setting gets changed
    {
        ...
        getEventDistributor().distributeEvent(SwitchRendererEvent(renderSettings.getRenderer());
        ...
    }

In other words, we pass the new renderer as information in the
SwitchRendererEvent and only update the `currentRenderer` variable when
we actually process that event.