SonarQube: fix incorrect use of std::condition_variable::wait()
The problem is that std::condition_variable::wait() can have spurious
wake ups. The solution is to pass an extra predicate parameter to
wait(). Though that predicate should then test for an extra
std::atomic<bool> variable, and we first need to set that bool before
calling std::condition_variable::notify_all(). Doable, but complex.

C++20 introduced a better suited synchronization primitive for this
problem: std::latch
    https://en.cppreference.com/w/cpp/thread/latch.html
That should be correct, and it also results in simpler code.

BIG CAVEAT: I don't have access to a windows computer with Midi
equipment, so I could not test this patch.