escape_newline: use c++23 resize_and_overwrite()
At least in escape_newline::decode(), there we know a cheap/tight upper
bound for the final string size.

There was also a comment to use it in escape_newline::encode() but there
the string can grow (up to 2x) so there using resize_and_overwrite() may
not be a good idea. (This patch only removes that comment.)

I've also added escape_newline::decode_inplace(). At the moment this
function isn't used yet (but I did extend the unittest for it). But if
we start using escape_newline more in the future, then it's good to
already have it (because it's a lot cheaper compared to the non-inplace
version).