Use c++23 resize_and_overwrite in strCat()
and strAppend().

This and the previous few patches use
std::string::resize_and_overwrite() (new in c++23), but I didn't explain
yet what that does:

Before we would have code like:
    std::string myString(size, ' '); // create a string containing 'size' number of spaces
    initialize(myString); // overwrite the content of 'myString'
The 'initialize()' function could e.g. be a C-style function (taking a
pointer to some buffer that will be initialized) or it could be a c++
function like strCat (as in this patch).

So we create a string with an internal buffer of the correct size, then
needlessly initialize that buffer, only to immediately after overwrite
that buffer with new contents.

This pattern can be "solved" via the new (c++23) resize_and_overwrite()
function. The above code then becomes something like:
    std::string myString;
    myString.resize_and_overwrite(size, [](char* buf, size_t max_capacity) {
        auto actualSize = initialize(buf, max_capacity);
        return actualSize; // actualSize is allowed to be smaller than max_capacity
    });
So this also allows to overestimate the required capacity and then only
after initialization report the actual size.

For more details see:
    https://en.cppreference.com/w/cpp/string/basic_string/resize_and_overwrite