Remove empty parameter list from lambda
When writing a c++ lambda expression it's allowed to omit an empty
parameter list (you don't have to write an empty pair of parenthesis).
Except that before c++23 there were some exception where you _did_ need
to write them. With c++23 the syntax is more regular.

I tested that gcc-12 and clang-17 (the minimum versions we require) do
already support this feature. Recent msvc versions should also support
this. Though I'm not sure that the version we have on the github action
runners is sufficiently recent. Let's just test, if not then we'll
revert this commit.