Adjust (relative) PSG volume in cartridges that have a PSG.
For details see #1934.

Summary: our PSG implementation has relatively low volume. Most devices use a
volume of 9000, but the PSG of 21000 in all machine configurations. If both PSG
and SCC are in a single devices, there is only one volume setting, but
apparently the relative volume between PSG and these devices must be
21000/9000, so let's adjust to that for now. Of course due to how the cartridge
is set up or programmed (FPGA) the volume may be different. But let's assume
it's similar to the other chips then and not like the relatively soft openMSX
default output volume until we have measurements to adjust it. So far only
Yamanooto and the HamarajaNight cartridges have been actually measured.