Properly 'sync' the VDPCmdEngine state before peeking its register values
For efficiency reasons we don't keep the VDPCmdEngine status constantly
up-to-date. That is we don't emulate this device cycle-by-cycle in sync
with all the other devices. Instead we try top emulate each device in
larger time chunks, and only force the values to be up-to-date when the
CPU (Z80) interacts with the device (I'm simplifying a bit).

This approach works fine if we do indeed synchronize at the appropriate
places. Though we forgot to synchronize on peeking the VDPCmdEngine
registers. (I checked, we do properly sync for e.g. the regular and the
status VDP registers).