hammingWindow: replace vector with MemBuffer
For the spectrum visualization (Tools > Audio channel viewer) we
calculate and cache hamming windows of different lengths. Before this
patch we stored those in a std::vector<float>.

This patch changes that to MemBuffer<float>. The reason is that resizing
a vector will also initialize all (new) elements. But immediately after
we'll overwrite all elements with a different value. There are
alternatives like "reserve and push_back" or
"to_vector(view::transform(...))". But since we already have the
MemBuffer utility, using that is even a tiny bit better than these
alternatives.