RomDatabase: tweak growing 'dumps'
Replace:
  dumps.resize(dumps.size() + 1);
with:
  dumps.emplace_back();

The former results in linearly growing the 'dumps' vector. The latter
results in geometric growth. Thus the former can be problematic when
called in a loop (resulting in O(N^2) behavior), the latter is fine
(loop remains O(N)).

In practice the 'dumps' vector never grows very large, so this
difference doesn't really matter. But it's good to fix anyway (and the
new code is also more compact).