Fix iterating over a changing vector
The method ImGuiManager::executeDelated() can be used to execute some
action at a later time, typically outside of the ImGui rendering stuff.
The problem was the sometimes executing such a delayed action queues
another delayed action. And in that case we iterate (while executing the
actions) over a changing vector (push_back() for queueing the next
action).