Fix(?) OPL3 4op stuff, issue#1525
    [BUG] Moonsound 4-op FM syntheses broken #1525
    https://github.com/openMSX/openMSX/issues/1525

Finally ... This is a very old issue, the bug was introduced +10 years
ago (though the issue was 'only' reported 2 years ago).

The difficulty with investigating this issue was that building such +10
year old openMSX versions isn't easy (requires old library version, old
python version, workarounds for code constructs that were technically
bugs but were accepted by the compilers at the time, ...). I once
started this, but because it was more difficult than expected I let it
rest for a while ... and then forgot to continue.

Yesterday I saw a reminder on MRC (post by GhostwriterP):
    https://www.msx.org/forum/msx-talk/general-discussion/attempt-at-automating-opl4-fm-configuration
And then I picked it up again.

This time, instead of trying to build old openMSX versions, I ported old
versions of the YMF262 code to the current openMSX version (also not a
trivial task, but I did succeed this time).

I found that the following commit was (likely) responsible:
    commit 1ff79a0848f2c206a23874525e7a90382848d937
    Author: Wouter Vermaelen <m9710797@users.sourceforge.net>
    Date:   Fri Aug 8 19:09:50 2008 +0000

    no need to test for zero, calculation yields the same result

In isolation this commit makes a transformation that _should_ be
correct. Though assuming the other (already existing) code is also
correct. And that assumption turned out to be wrong:
* The variable 'Incr' is a pre-calculation based on 'ch.fc', 'mul',
  'block_fnum', ... But from what I could tell these values were not
  consistent.
* So this commit uncovered bugs that were already present, but didn't
  manifest itself.
Yesterday I tried to fix this but I was not successful.

This morning I discussed the problem with 'ValleyBell'. He was able to
reproduce the bug in the MAME code (after making a similar change as in
1ff79a0848f2c). And importantly he also understood what went wrong and
was able to fix it. Both in the MAME and in the openMSX version of the
code.

So thanks a lot to 'ValleyBell'!!