Improve (fix?) clipping at top/bottom border for LINE command
Thanks to this issue reported by 'hra1129':
    VDP Command LINE issue #1961
    https://github.com/openMSX/openMSX/issues/1961

When a LINE command (going downwards) moves beyond the bottom border, it
wraps around and continue from the top. But when a command (going
upwards) hits the top border, the command stops.

We have the same behavior for block commands, only for the LINE command
this (surprising?) behavior wasn't implemented yet. See also the
comments in the code.