m9710797 Contribution with huge changes to improve code overall and enable Windows compilatim9710797u on. Changes by m9710797:
Here are some highlights of the changes I made:

Could it be you made a mistake in your last update, and accidentally changed IO port 0x5a to 0x57? I changed that back to 0x5a, openMSX should match the existing hardware.
No more #include of platform specific files. All is handled via our Socket.hh abstraction (for the network stuff) and via Timer.hh (for the sleep() stuff).
In fact most of the #ifdef _WIN32 stuff is gone. (Almost) all differences between Linux and windows should be hidden inside Socket.hh and Timer.hh.
I renamed serverThread to readLoop. The stuff in MSXPiDevice is not really a server.
I created a helper function close() to close the socket and set the socket-variable to the invalid value. Because this sequence was repeated several time.
I've renamed sockfd to sock (socket would be a better name, but that conflicts with the function with this name). because on Linux it's indeed a file descriptor. But on Windows it's a SOCKET handle. Our file Socket.hh also hides these differences. In our code we can just use SOCKET as the type, and don't worry about the platform differences.
I've removed txQueue. Now that we handle sending from the main thread this queue is no longer needed.
I've simplified the implementation of readIO() and peekIO() a bit. Handling of port 0x56 was identical between both, so I don't repeat that code. I tried to take the lock (mtx) for smaller sections in the code, only when really doing stuff with rxQueue. Because readIO() and peekIO() are so closely related, I've moved them closer together in the code.
Before we had both the variables serverAvailable and sockfd (now renamed to sock). But these two both indicated the same thing: serverAvailable should always be equal to sock != OPENMSX_INVALID_SOCKET. So I removed serverAvailable.
Ah, I did make one functional change on reading IO port 0x5a: before when no data was available in rxQueue the variable readRequested was NOT set to false. Now it's always set to false after reading port 0x5a. I thought that was more logical, but please carefully review this change. Is my change correct?
I've made similar simplifications in writeIO().
I've made major changes in the helper thread (in `readLoop()):
The windows socket startup/cleanup code is all handled via SocketActivator (part of Socket.hh). This also takes care of not initializing this subsystem twice when e.g. some other part of openMSX already initialized it.
I no longer use inet_pton() to parse the fixed string 127.0.0.1 instead I directly initialize the address with INADDR_LOOPBACK. I also removed the member variables serverIP and serverPort. If in the future we want to make the port configurable(??) we may want to re-add this. (I don't think we should make the IP-address configurable).
You use select() to be able to specify a timeout. But in a way that's still busy-looping. I replace that with a blocking read call. And to abstract some minor differences between Linux and Window, we replace recv() with sock_recv() from our Socket.hh file.
In the previous point I say we do a blocking read, that's only partially true. On Windows it's true, there sock_recv() really blocks until there is data available. On Linux we instead use another openMSX utility Poller. We call poller.pol(), that one blocks until there is data available, and only then we read that data via sock_recv().
The reason for this Window/Linux difference is related to stopping the thread. On Windows when the main thread calls sock_close() the blocking call to sock_revc() does return with an error. In other words: it unblocks that call. On Linux (or more general on Unix platforms) that behavior is not guaranteed, and there we need a more complex mechanism. But that's all hidden in the poller.poll() call. That one does get unblocked when the main thread calls poller.abort().
I left one TODO in the code: after we've received our data via sock_recv() we put that data in rxQueue. But currently that queue can grow unbounded in size. Probably we should give it an upper limit, so that the python server cannot arbitrarily keep on filling memory in the openMSX process.
I've removed the variables statusReg, controlReg and dataReg. They were not used.
I've renamed running to shouldStop. That better reflects the intend of this variable (request the helper thread to stop).
I've changed the rxQueue type from std::queue to cb_queue. This is an openMSX specific class (so CoPilot likely doesn't know about it). It's a queue based on a circular buffer, it's a bit faster than std::queue (which is internally based on a std::deque), and in a few cases it offers a more convenient interface (e.g. it offers clear() and a pop_front() which in one call both returns the front value and removes it from the queue).