Greatly speedup reading blocks from 'slotted memory'
"Slotted memory" is the 64kB Z80 address space, but for the 4 primary and
4 secundary slots. Thus it's 4 x 4 x 64kB = 1MB in size.

The debugger hex editor sometimes takes a snapshot of the full
debuggable. This is used to be able to highlight changes (right-click
context menu, highlight changes).

But peeking "Slotted memory" was fairly slow. It performs +1 million
individual byte reads, each with some checks and lookups in the slot
structure.

In this patch we speed that up by using the 'readCacheLine' mechanism.
That is for ROM/RAM like regions (that's by far most of the address
space), we now process the request in blocks of 256 bytes.

In total I measured a factor between 10x and 20x for making a full
snapshot of "Slotted memory".

--

The trigger for investigating this was:
    MSX screen is not updated periodically when stepping through in the debugger #1962
    https://github.com/openMSX/openMSX/issues/1962

After a lot of detective work I found out that the trigger for this
issue was having the "slotted memory" hex editor open.

I do not yet understand why it's not just slow, and why it also drops
animation frames (both {min,max}frameskip=0). That's something to
investigate later.