DiskManipulator: Always set BPB hidden sectors to 0 (issue .
DiskManipulator formats 1DD and 2DD floppies with the BPB hidden sectors field
set to 1, however disk images created with _FORMAT always have a hidden sectors
count of 0. This matches dos1BootBlock and dos2BootBlock in BootBlocks.cc.

It is also never set for larger HD images by either Nextor FDISK or Sunrise
IDEFDISK 3.1. In the latter case, not even for 32MB partitions, for which
DiskManipulator even sets the hidden sector count to 16.

According to Wikipedia it should only be specified if the volume is preceded by
a partition table, but this seems unused on MSX (as noted in the MSXBootSector
struct). If we really wanted to set it, it should be done by the partitioning
code. But I don’t think we do.

This change just sets it to 0 at all times.

Closes #1976.