clang-tidy: fix all 'llvm-include-order' warnings
We already tried to following this rule in openMSX, but (before I found
this clang-tidy check) had no way to verify it.

For a detailed description see here:
    https://clang.llvm.org/extra/clang-tidy/checks/llvm/include-order.html
    https://llvm.org/docs/CodingStandards.html#include-style

I'll give a summary using an example, let's take the file
`src/memory/MegaFlashRomSCCPlusSD.cc`, it lists the following #includes:

    #include "MegaFlashRomSCCPlusSD.hh"

    #include "CheckedRam.hh"
    #include "SdCard.hh"

    #include "CacheLine.hh"
    #include "DummyAY8910Periphery.hh"
    #include "MSXCPUInterface.hh"
    #include "serialize.hh"

    #include "enumerate.hh"
    #include "narrow.hh"
    #include "ranges.hh"
    #include "xrange.hh"

    #include <array>
    #include <memory>

* First list the `.hh` file that belongs to this `.cc` file
  (`MegaFlashRomSCCPlusSD.hh` in this case). This is a single line, the
  groups below can contain more than one #include.
* Followed by all the header files in the same directory (`src/memory`
  in this case).
* Followed by other openMSX specific header files (we don't try to split
  these up further over directories).
* Followed by header files from `src/utils` (these are more or less
  independent from the rest of openMSX).
* And finally system header files.

Within each group the lines are sorted (according to the `sort` tool
with default options, so in ASCII order).

The groups are ordered from more specific to more generic. This, more or
less, implies the header files in higher groups may themselves #include
headers from a lower group, but not the other way around.

The reasoning behind this is:
* This group ordering makes it more likely to detect 'incomplete' header
  files errors. Meaning a header that uses some symbol without first
  #including the corresponding header itself.
* Having the items sorted makes it easier to spot duplicate #includes.

Even though we already tried to follow this rule in openMSX, this patch
is relatively large. It mostly just reorders #includes, or adds some
blank lines. Though in some cases it did also already allow to remove
some unnecessary #includes.