Add 'extra' parameter in File::mmap()
Allow to request some extra bytes after a mapped file.

In several places in openMSX we have this pattern:
* Query the filesize.
* Allocate a buffer large enough for that size + some extra bytes.
* Read the file in the buffer.
* Fill the extra bytes with zero (or something else).
A typical use for this is:
* Parsing a file, but have the guarantee that it is zero-terminated
* Or have the guarantee that there are still 8 valid bytes after the
  file. That allows to do parsing with 64-bit reads, without the risk of
  reading past valid memory (and triggering an MMU protection error).

When mmap()ing a file, the OS handles that with page granularity
(typically a page is 4096 bytes). Very often a file is not an exact
multiple of a page, and then the left-over bytes in the last page are
already filled with zeros. In this patch we exploit this, so we avoid
having to manually allocate a buffer, and read the full file. And
important: wait for that file read to finish. In other words: we retain
the advantage of asynchronous I/O (we can start parsing the start of the
file while the tail of the file is still being read from disk).

More in detail, the strategy we use is:
1) If there are enough tail bytes in the last mmap()ed page, then we're
  done.
2) (Not on windows) otherwise we try to map an anonymous page right after
  the mapped page.
3) If step 2 failed, we fall back to allocating a buffer.

This patch adds to new infrastructure, but does not yet use it. That's
for the next patch.