Don't use mmap() for '.filecache'
We got a crash (bus error) triggered by this scenario:
* We mmap() the file '.filecache' in memory.
* While parsing this file we keep pointers to that data (e.g. to avoid
  making copies, or to only lazily parse the date/time-string into a
  'time_t').
* So we keep the mmap() around even after parsing.
* When we exit openMSX, and when the file-cache was changed (more
  or updated entries) we write that file back to disk. So we open and
  truncate the '.filecache' file.
* During the write we iterate over the entries, BUT this still refers to
  memory of the original mmap()ed file. And we've just truncated that
  backing file.
* This can cause crashes (bus error).

To fix this we revert part of edbdbe0c42cabfd (I did verify that the
other parts of that commit are safe). So instead of mmap() we read the
whole file in a buffer (we don't mmap() it anymore).

On unix a better solution could have been:
* Write to a temporary file.
* Then rename that file to '.filecache'.
Though unfortunately that doesn't work on windows.