mmap() .rom files as 'non-const'
To allow for IPS-patches. In the c++ code we lie and say this is 'const'
data. But then later for IPS patches we cast away this 'const'. That's
fine when the underlying memory is in fact also writable. But if that
memory was really read-only (because it's mmap()ed as read-only) then
trying to write triggers an MMU protection error.

Got recently broken in commit 185ec2899b865978. At least it shows that
mapping read-only does work ;-)