Correctly initialize MSXCPU::slots[] on loadstate (e.g. for reverse)
This was responsible for
    Reverse issue on Spectravideo SVI-328
    https://github.com/openMSX/openMSX/issues/2002

This issue is about Spectravideo. But maybe it could also have gone
wrong on MSX (though we never(?) saw it there??).

In the end this caused corrupt cacheline administration (and then reads
from one slot were wrongly using cache lines from another slot). Was
tricky to debug, but once identified it was easy to fix.