Allow to double click on an instruction in the disassmbly view
To scroll the disassmbly view to the address used in the instruction.
For example instructions like:
    call #1234
    jp nz,#1234
    ld hl,(#1234)
If you double click those, the disassembly view scrolls to address
0x1234.

We already had something like this, but for some reason we only made it
work if the address had an associated symbol.

Maybe this was requested in
    https://github.com/openMSX/openMSX/issues/2006
Description in the ticket isn't very clear (to me).