Detect invalid external slot specification
E.g. our CIEL_Expert-Turbo machine config has:
    ...
    <primary slot="2">
        <MemoryMapper id="Main RAM">
            ...
        </MemoryMapper>
    </primary>

    <primary external="true" slot="2"/>

So it both inserts devices in slot 2 and marks that slot as external.

That's wrong. But before this patch we silently accepted it. Though
later on it could cause inconsistencies. E.g. in the cartridge
visualization code (the ImGui setup viewer) we don't expect to find a
"machine" config in an external cartridge slot.

Now we refuse to load the invalid CIEL_Expert-Turbo machine config.
We'll fix that config in the next commit. Unfortunately this break
savestates with that machine. Is that a big problem? Is it a commonly
used machine?