Cleanup stale SocketConnection objects
Each time an external process connects to openMSX, disconnects, and
re-connects, openMSX would leave a zombie thread behind. That is the
thread is no longer running, but the parent thread would not "join" the
thread. When this happens too many times (hundreds) we eventually run
out of resources.

Technically it was not a 'real' leak. At the end when openMSX itself
would stop it _would_ properly clean up the threads (I mean openMSX
cleans up, not the OS). But of course that's not good enough in the
above scenario.

So what we do now is at certain moments explicitly check for stale
connections (those that are closed from the client side) and clean them
up. We do that right before starting a new connection.

This may not be the cleanest solution (we don't cleanup at the earliest
possible moment), but it's good enough for now (and was the easiest to
implement). A cleaner design would be to not use a separate thread for
each connection, but handle all connections in only one thread (possibly
even the main thread). Though that's a bit harder to do in a portable
way. Let's wait with that until there's a good library solution for that
(preferably the c++ standard library so that we don't need extra
dependencies).