Filename tab-completion usability improvements
Inspired by:
    https://github.com/openMSX/openMSX/pull/2017
More specifically this part:
<quote>
    Possible strings (candidates) now only contain last component of the path.
        Previously, all items have common path which typed in console.
</quote>

This makes the candidates shown in the (new) pop-up less redundant, and
thus more useful items fit in a (smaller) pop-up.

---

Internally the tab-completion functions are changed to work with
'CompletionCandidate' (a small struct) instead of strings. With this
new struct it's possible to have a different 'display' string vs
'replacement' string. And it's possible to indicate if the token is
complete or if it's only a partial completion (e.g. complete the
directory portion in a longer filename).

This makes the completion code more general. The string-completion and
the filename-completion code now share (more) of this general code.