Greatly simplify LocalFile::preCacheFile()
Via posix_fadvise() instead of manually doing reads in a separate
thread.

I remember the original motivation for pre-caching was using dsk images
from a CDROM. After some inactivity the CDROM spins down and then when
doing a new read from the dsk file, openMSX would hang for a few seconds
because first the CDROM had to spin up again. To mitigate that we would
try to read the whole file in the OS disk cache.

The above use case is probably no longer relevant. So it's good that we
can greatly simplify the code to support it. On Unix it became basically
a one liner. And it's not too bad that on Windows it doesn't do anything
anymore.