Raster beam viewer: fixed wrong assertion
Within a frame, there cannot be a single line that has mixed width:
screen-mode splits are postponed till the next line. But in the raster
beam viewer we mix two consecutive frames, and there it can easily
happen that from one frame to the next the same line number has
different width. So the 'assert(wl == wl)' was just wrong.