ResampleHQ: fixed out-of-bound read (off-by-one)
In 'ResampleCoeffs.ii' we define a filter with (half-)length 2463
coefficients. For interpolation between coefficients we pad this array
with one extra zero. But in some place in the code we then wrongly took
that to mean a filter with 2664 coefficients (one too many). And if we
then actually need that last coefficient, and interpolate it, we read 1
past the end of this array.

This triggered purely as a coincidence while creating a HQ-resampler for
the AY8910 when running at speed=0.1