Tweak Pluggable/Connector, string_view -> zstring_view
The methods
  getName(), getDescription(), getClass()
in the classes
  Pluggable, Connector
used to return a std::string_view. This patch changes that to
zstring_view.

We did already always return a zero-terminated string. But we 'lost'
that information in the type-system.

Now that we retain that information, we can optimize the
pluggableToGuiString() function in ImGuiConnector: no need to convert
std::string_view into std::string, only to add a zero-terminator
(because now we know it is already present).