Auto size column widths, whilst still keeping them in sync.
Patch created by Wouter. This makes the QSE look a bit more compact.

As Wouter explained to me:
We're currently using a "default table" (I mean, without much configuration).
But then you get a table where all the columns are the same width. And that
leaves quite a bit of empty space between the two columns. Dear ImGui has an
option to automatically determine the column width. That works, but it didn't
look very nice because then the four tables wouldn't be properly aligned.

I remembered that the standard Dear ImGui demo application has a solution for
this. It demonstrates tables that are "in sync": "Dear ImGui Demo > Tables &
Columns > Synced Instances" try resizing a column there; you'll see that the
other tables stay in sync. The way to do that is apparently to give all the
tables the same name.

It seemed simple, but the TreeNode is the game-changer. Our tables are each
under a different TreeNode. Because of that, the complete hierarchical name is
different for our four tables. After some tinkering, I created a new utility
function: "TreeNodeWithoutID()." It works just like the regular "TreeNode()"
function, except that (conceptually) it doesn't add an extra ID to the ImGui ID
stack. So, in terms of hierarchical name, the function is "transparent." And
this one works.

The "Quick Setup Editor" window is now a bit more compact. This creates more
usable space for long machine or media names.