Simplify showSetupOverviewMachine()
Don't handle findMachineInfo() returning nullptr. Can only happen when
the user change the openMSX installation while openMSX is running. It's
sufficient that we don't crash if that happens, but it's not needed to
try to properly show it in the GUI.

Handle special case before calling showSetupOverviewMachine(), possibly
by not calling this function at all. This removes the need to pass the
'viewMode' parameter.