Allow to specify a segment number of symbol files
We could already set an (optional) slot, now we can also set an
(optional) segment number.

Enter a value, or - to ignore (the latter is the default).

Reminder: for symbol files that contain symbol values > 0xffff, the
upper 16 bits are consider to be a segment number. In other words, such
files already do specify a segment per symbol, and then we don't allow
to set the segment for the file.

This should implement bengalack's request from this issue:
    https://github.com/openMSX/openMSX/issues/1702#issuecomment-3497322748