Fix compilation error on (older?) visual studio 2022
Thanks a lot to Suzumizaki for reporting:
    Fixed an issue where bbc60d81 prevented Visual Studio 2022 from building. #2046
    https://github.com/openMSX/openMSX/pull/2046

The problem was that 'constexpr std::string' is not (yet) supported on
(older?) versions of Visual Studio. (The code compiled fine on github,
also with vs2022, possibly a newer version).

The proposed patch was to change 'constexpr' to 'const'. But I fixed it
in a different way: we didn't really need a 'std::string', a
'std::string_view' was sufficient (and that one can be 'constexpr').