Fix issue #2050 Color address not shown correctly in debugger tile view
See:
    https://github.com/openMSX/openMSX/issues/2050

This got broken in revision 6174c91897ab while implementing
    https://github.com/openMSX/openMSX/issues/2044
Internally the tile viewer then kept two color tables (I mean the
'VramTable' class that implements the VDP table address calculations),
one used for rendering (which can be overridden) and one to show the
color address. Though because of that we no longer correctly called
'setIndex' on that 2nd table, and because of that the address was shown
as-if you always selected address 0x00000 for the color table. The
rendering was still correct because that uses a different table.

I fixed it by going back to a single table. A consequence is that now
when the 'override color tabel' option is active, the color address is
always shown as 0x0000-0x0007. But that's probably OK.