Added 'TraceValue' class
This is similar to
    std::variant<std::monostate, uint64_t, double, std::string>
but optimized for memory size:
* sizeof(TraceValue) == 16
* sizeof(std::variant<..>) > 16

Strings up to 15 bytes are stored in TraceValue itself, longer strings
require heap allocation. Though as a tradeoff strings cannot store
embedded 0-chars.

Next to 'std::variant<..>' we could also have used 'TclObject', though
that's even larger in size.

The main purpose of this class is to store many (thousands or even
millions) or these objects without consuming too much memory. It will be
used as the store for the trace data (see future patch).