change Debugger::probes
From an (unordered) hash_set to a vector sorted on ProbeBase::getName().

This mean insert/remove/lookup changes from O(1) to O(log N). But we
only have a handful of probes, so this should be fine (and it's anyway
not a common operation).

The main reason for this is that it avoids a sort (on each frame) in the
GUI, in a later patch.