Fix text placement in ImGuiTraceViewer drawEventsValue()
This is a regression introduced late during development of this feature
(prior to pushing). Because of numerical inaccuracies when zoomed in
very far, I had to change how horizontal scrolling worked. This bug is
because I forgot to update this code part.

Thanks to 'max_iwamoto' for trying out the example and reporting the
problem.